+++ src/subtract.cpp
@@ -1,7 +1,7 @@
 #include "subtract.h"
 
 namespace Calculator{
-    float subtract(float a, float b)
+    double subtract(double a, double b)
     {
         return a - b;
     }
